fix: move utils
