+++ src/modules/v1/schedule/schedule.controller.ts
@@ -2,10 +2,10 @@ import { createRoute, z } from "@hono/zod-openapi"
 import { asc, eq } from "drizzle-orm"
 import { scheduleTable, Schedule } from "../../../db/schema"
 import { createAPI } from "../../api"
-import { buildResponseSchemas } from "../../utils/response"
+import { buildResponseSchemas } from "../../../utils/response"
 import { scheduleResponseSchema } from "./schedule.schema"
 import { Cache } from "../cache"
-import { getSecsToMidnight } from "../uitls"
+import { getSecsToMidnight } from "../../../utils/time"
 
 const api = createAPI()
